refactor(client): dedupe day-timeline fire-time math into lib/schedule
pages/routines/DayTimeline.tsx reimplemented the cron-to-fire-times logic
(including the midnight-boundary seed trick) locally and untested, instead of
reusing the equivalent, already-tested fireTimesOnDay already exported for
this purpose from heatmap/dayTimelineMath.ts. Moved fireTimesOnDay into
lib/schedule.ts (its natural shared home, alongside occurrencesPerDay) so
both day-timeline views resolve "what fires today" from one tested
implementation instead of two that can silently drift apart.
